hematology-paper-writer/requirements.txt
# Hematology Paper Writer Dependencies

# Core dependencies
pyyaml>=6.0
python-docx>=0.8.11
httpx>=0.24.0
xmltodict>=0.13.0
tqdm>=4.65.0
requests>=2.31.0

# Web UI dependencies (Week 0)
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
Pillow>=10.0.0
